+++ FirebaseAI/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+- [removed] Removed deprecated Imagen APIs. Note that the Imagen service turned down on August 17, 2026.
 - [feature] Added support for `RealtimeInputConfig` in `LiveGenerationConfig`. (#16441)
 - [feature] Added support for `sendStartActivityRealtime` and `sendStopActivityRealtime`
   in `LiveSession`. (#16441)